+++ collections/ansible_collections/deamen/general/CHANGELOG.md
@@ -5,5 +5,7 @@ All notable changes to this collection will be documented in this file.
 The format is based on Keep a Changelog (https://keepachangelog.com/en/1.0.0/) and this project follows Semantic Versioning.
 
 ## Unreleased
+- Add check_is_container module
+- Add install_dev_packages role
 - Add install_system_ansible role
 - Initial scaffold